Was having huge issues trying to figure out AnimatedContent, was trying to transform from MenuOptionButton to their respective option screen. Learned that transforming from Button Composables does not work, the Button just moves according to Transitions rather than morph into end result. The better way would be to make a button from a Box Composable, which was done with Fab Composable to test it out. It looks great now! SolitaireMenu renamed to StatsScreen and moved to its own file. Each Menu option will have its own screen/file/Composable. Will remove Scaffold from MainActivity as it is more of a nuisance since top/bottom bar have to be animated away when a Menu option screen is shown.
